report_service/requirements.txt
azure-functions
azure-eventgrid
azure-functions-durable
azure-data-tables
pyodbc
requests